cleaned up some code
