--- frameRateDisplay.py
@@ -1,4 +1,0 @@
-import bge.logic as logic
-cont = logic.getCurrentController()
-own = cont.owner
-own['Text'] = logic.getAverageFrameRate()
